Change version

+++ PyOphidia/utils/run.py
@@ -23,8 +23,8 @@ import cwltool, cwltool.factory
 from PyOphidia import Workflow, Experiment, client
 
 parser = argparse.ArgumentParser()
-parser.add_argument('name', type=str, help='Workflow file name')
-parser.add_argument('--args', type=str, help='Workflow arguments')
+parser.add_argument('name', type = str, help = 'Workflow file name')
+parser.add_argument('--args', type = str, help = 'Workflow arguments')
 args = parser.parse_args()
 
 cwl_args = {}
@@ -49,12 +49,13 @@ json_request = result["outputexperiment"]["location"][7:]
 #    print(f.read())
 
 e1 = Experiment.load(json_request)
-#e1.check()
+if not e1.check():
+    raise Exception("Experiment is not valid")
 
 os.remove(json_request)
 os.rmdir(json_request.rsplit('/', 1)[0])
 
-ophclient = client.Client(read_env=True)
+ophclient = client.Client(read_env = True)
 Workflow.setclient(ophclient)
 
 w1 = Workflow(e1)
